authsolbackend/requirements.txt
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
construct==2.10.68
construct-typing==0.5.6
exceptiongroup==1.2.2
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonalias==0.1.1
MarkupSafe==3.0.2
pycryptodome==3.22.0
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
typing_extensions==4.13.0
websockets==15.0
Werkzeug==3.1.3
zipp==3.21.0
